# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type CitizenProfileRole = typeof CitizenProfileRole[keyof typeof CitizenProfileRole];


export const CitizenProfileRole = {
  citizen: 'citizen',
  admin: 'admin',
} as const;

export interface CitizenProfile {
  id: number;
  clerkUserId: string;
  fullName: string;
  aadhaarNumber: string;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  role: CitizenProfileRole;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileSync {
  fullName: string;
  /** @nullable */
  email?: string | null;
}

export type UpdateRequestFieldName = typeof UpdateRequestFieldName[keyof typeof UpdateRequestFieldName];


export const UpdateRequestFieldName = {
  mobile_number: 'mobile_number',
  email: 'email',
  address: 'address',
} as const;

export type UpdateRequestStatus = typeof UpdateRequestStatus[keyof typeof UpdateRequestStatus];


export const UpdateRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface UpdateRequest {
  id: number;
  userId: number;
  /** @nullable */
  citizenName?: string | null;
  fieldName: UpdateRequestFieldName;
  /** @nullable */
  oldValue?: string | null;
  newValue: string;
  status: UpdateRequestStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type UpdateRequestInputFieldName = typeof UpdateRequestInputFieldName[keyof typeof UpdateRequestInputFieldName];


export const UpdateRequestInputFieldName = {
  mobile_number: 'mobile_number',
  email: 'email',
  address: 'address',
} as const;

export interface UpdateRequestInput {
  fieldName: UpdateRequestInputFieldName;
  /** @minLength 1 */
  newValue: string;
  /** @nullable */
  otpToken?: string | null;
}

export interface AdminDecisionInput {
  /** @nullable */
  adminNote?: string | null;
}

export interface AdminRequestsPage {
  requests: UpdateRequest[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminCitizensPage {
  citizens: CitizenProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminStats {
  totalCitizens: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  totalRequests: number;
  recentActivity: UpdateRequest[];
}

export interface AuditLog {
  id: number;
  action: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  adminId?: string | null;
  /** @nullable */
  targetField?: string | null;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface AuditLogsPage {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export type OtpSendInputType = typeof OtpSendInputType[keyof typeof OtpSendInputType];


export const OtpSendInputType = {
  mobile: 'mobile',
  email: 'email',
} as const;

export interface OtpSendInput {
  type: OtpSendInputType;
  value: string;
}

export interface OtpSendResult {
  success: boolean;
  message: string;
  expiresIn: number;
}

export type OtpVerifyInputType = typeof OtpVerifyInputType[keyof typeof OtpVerifyInputType];


export const OtpVerifyInputType = {
  mobile: 'mobile',
  email: 'email',
} as const;

export interface OtpVerifyInput {
  type: OtpVerifyInputType;
  value: string;
  code: string;
}

export interface OtpVerifyResult {
  valid: boolean;
  /** @nullable */
  token?: string | null;
  message: string;
}

export interface ChatbotQuestion {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface DocumentSource {
  sourceName: string;
  excerpt: string;
}

export interface ChatbotAnswer {
  answer: string;
  sources: DocumentSource[];
  conversationId: string;
  retrievedCount: number;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  sources: DocumentSource[];
  createdAt: string;
}

export interface CitizenDashboardStats {
  totalRequests: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  recentRequests: UpdateRequest[];
}

export type AdminListRequestsParams = {
status?: AdminListRequestsStatus;
search?: string;
page?: number;
limit?: number;
};

export type AdminListRequestsStatus = typeof AdminListRequestsStatus[keyof typeof AdminListRequestsStatus];


export const AdminListRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export type AdminListCitizensParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListAuditLogsParams = {
page?: number;
limit?: number;
userId?: string;
};