# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Government Citizen Service Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/requests:
    get:
      operationId: adminListRequests
      tags: [admin]
      summary: List all citizen update requests (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of update requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRequestsPage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/requests/{id}/approve:
    post:
      operationId: adminApproveRequest
      tags: [admin]
      summary: Approve an update request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDecisionInput"
      responses:
        "200":
          description: Approved request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/requests/{id}/reject:
    post:
      operationId: adminRejectRequest
      tags: [admin]
      summary: Reject an update request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDecisionInput"
      responses:
        "200":
          description: Rejected request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/citizens:
    get:
      operationId: adminListCitizens
      tags: [admin]
      summary: List all citizens (admin only)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of citizens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCitizensPage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit logs (admin only)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogsPage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /chatbot/ask:
    post:
      operationId: chatbotAsk
      tags: [chatbot]
      summary: Ask the AI chatbot a question (RAG-based)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotQuestion"
      responses:
        "200":
          description: AI answer with source references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotAnswer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /chatbot/history:
    get:
      operationId: getChatHistory
      tags: [chatbot]
      summary: Get chatbot conversation history for current user
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: clearChatHistory
      tags: [chatbot]
      summary: Clear chatbot conversation history for current user
      responses:
        "204":
          description: History cleared
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CitizenProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        fullName:
          type: string
        aadhaarNumber:
          type: string
        mobileNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
          enum: [citizen, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - fullName
        - aadhaarNumber
        - role
        - createdAt
        - updatedAt
...
    UpdateRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        citizenName:
          type: ["string", "null"]
        fieldName:
          type: string
          enum: [mobile_number, email, address]
        oldValue:
          type: ["string", "null"]
        newValue:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - fieldName
        - newValue
        - status
        - createdAt
        - updatedAt
...
    AdminDecisionInput:
      type: object
      properties:
        adminNote:
          type: ["string", "null"]
    AdminRequestsPage:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/UpdateRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - requests
        - total
        - page
        - limit
    AdminCitizensPage:
      type: object
      properties:
        citizens:
          type: array
          items:
            $ref: "#/components/schemas/CitizenProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - citizens
        - total
        - page
        - limit
    AdminStats:
      type: object
      properties:
        totalCitizens:
          type: integer
        pendingRequests:
          type: integer
        approvedRequests:
          type: integer
        rejectedRequests:
          type: integer
        totalRequests:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/UpdateRequest"
      required:
        - totalCitizens
        - pendingRequests
        - approvedRequests
        - rejectedRequests
        - totalRequests
        - recentActivity
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        userId:
          type: ["integer", "null"]
        adminId:
          type: ["string", "null"]
        targetField:
          type: ["string", "null"]
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - createdAt
    AuditLogsPage:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - logs
        - total
        - page
        - limit
...
    ChatbotQuestion:
      type: object
      properties:
        question:
          type: string
          minLength: 1
        conversationId:
          type: ["string", "null"]
      required:
        - question
    ChatbotAnswer:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSource"
        conversationId:
          type: string
        retrievedCount:
          type: integer
      required:
        - answer
        - sources
        - conversationId
        - retrievedCount
    DocumentSource:
      type: object
      properties:
        sourceName:
          type: string
        excerpt:
          type: string
      required:
        - sourceName
        - excerpt
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSource"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - sources
        - createdAt
...