# lib/api-zod/src/generated/types/adminRequestsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateRequest } from './updateRequest';

export interface AdminRequestsPage {
  requests: UpdateRequest[];
  total: number;
  page: number;
  limit: number;
}