# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  action: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  adminId?: string | null;
  /** @nullable */
  targetField?: string | null;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  /** @nullable */
  details?: string | null;
  createdAt: Date;
}