# lib/api-zod/src/generated/types/chatbotAnswer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentSource } from './documentSource';

export interface ChatbotAnswer {
  answer: string;
  sources: DocumentSource[];
  conversationId: string;
  retrievedCount: number;
}