# lib/api-zod/src/generated/types/chatbotQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatbotQuestion {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  conversationId?: string | null;
}