# lib/api-zod/src/generated/types/citizenProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Government Citizen Service Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { CitizenProfileRole } from './citizenProfileRole';

export interface CitizenProfile {
  id: number;
  clerkUserId: string;
  fullName: string;
  aadhaarNumber: string;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  role: CitizenProfileRole;
  createdAt: Date;
  updatedAt: Date;
}